Move load_conf function
