+++ plugins/modules/mongodb_replicaset.py
@@ -220,6 +220,7 @@ from ansible.module_utils.six import binary_type, text_type
 from ansible.module_utils.six.moves import configparser
 from ansible.module_utils._text import to_native
 from ansible_collections.community.mongodb.plugins.module_utils.mongodb_common import check_compatibility
+from ansible_collections.community.mongodb.plugins.module_utils.mongodb_common import load_mongocnf
 
 
 def replicaset_find(client):
@@ -300,23 +301,6 @@ def replicaset_remove(module, client, replica_set):
     #    db.remove_user(replica_set)
     # else:
     #    module.exit_json(changed=False, user=user)
-
-
-def load_mongocnf():
-    config = configparser.RawConfigParser()
-    mongocnf = os.path.expanduser('~/.mongodb.cnf')
-
-    try:
-        config.readfp(open(mongocnf))
-    except (configparser.NoOptionError, IOError):
-        return False
-
-    creds = dict(
-        user=config.get('client', 'user'),
-        password=config.get('client', 'pass')
-    )
-
-    return creds
 
 
 # =========================================
