+++ plugins/modules/mongodb_user.py
@@ -215,6 +215,7 @@ from ansible.module_utils.six import binary_type, text_type
 from ansible.module_utils.six.moves import configparser
 from ansible.module_utils._text import to_native
 from ansible_collections.community.mongodb.plugins.module_utils.mongodb_common import check_compatibility
+from ansible_collections.community.mongodb.plugins.module_utils.mongodb_common import load_mongocnf
 
 
 def user_find(client, user, db_name):
@@ -259,22 +260,6 @@ def user_remove(module, client, db_name, user):
         db.remove_user(user)
     else:
         module.exit_json(changed=False, user=user)
-
-
-def load_mongocnf():
-    config = configparser.RawConfigParser()
-    mongocnf = os.path.expanduser('~/.mongodb.cnf')
-
-    try:
-        config.readfp(open(mongocnf))
-        creds = dict(
-            user=config.get('client', 'user'),
-            password=config.get('client', 'pass')
-        )
-    except (configparser.NoOptionError, IOError):
-        return False
-
-    return creds
 
 
 def check_if_roles_changed(uinfo, roles, db_name):
